+++ base/deprecated.jl
@@ -1444,6 +1444,16 @@ end
 # PR #23332
 @deprecate ^(x, p::Integer) Base.power_by_squaring(x,p)
 
+# Deprecate defaulting to broadcast unknown values as scalars
+function Broadcast.collect_unknown(::Broadcast.Unknown, x)
+    depwarn("treating x::$(typeof(x)) as a scalar in broadcast is deprecated, use `Ref(x)` instead", (:broadcast, :broadcast!))
+    return Ref(x)
+end
+function Broadcast.collect_unknown(::Broadcast.Unknown, x::Union{AbstractArray,Ref})
+    depwarn("treating x::$(typeof(x)) as a scalar in broadcast is deprecated, use `fill(x)` instead", (:broadcast, :broadcast!))
+    return fill(x)
+end
+
 # Issue #25979
 # The `remove_destination` keyword to `cp`, `mv`, and the unexported `cptree` has been
 # renamed to `force`. To remove this deprecation, remove the `remove_destination` keyword